Review comments: - Extract CommitLogMarkers into standalone classes - Use CdcEventBuilder in CdcEvent constructor - Comment typo corrections
